API-903: Rename query/handler
