+++ src/Akeneo/Apps/back/Audit/Application/Query/CountDailyEventsByAppQuery.php
@@ -9,7 +9,7 @@ namespace Akeneo\Apps\Audit\Application\Query;
  * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
  * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class FetchAppsEventCountByEventQuery
+class CountDailyEventsByAppQuery
 {
     /** @var string */
     private $eventType;
