+++ src/Akeneo/Apps/back/Infrastructure/Symfony/Resources/config/handlers.yml
@@ -37,7 +37,7 @@ services:
             - '@akeneo_app.persistence.repository.app'
             - '@akeneo_app.service.client.regenerate_client_secret'
 
-    akeneo_app.audit.application.handler.fetch_apps_event_count_by_event:
-        class: 'Akeneo\Apps\Audit\Application\Query\FetchAppsEventCountByEventHandler'
+    akeneo_app.audit.application.handler.count_daily_events_by_app:
+        class: 'Akeneo\Apps\Audit\Application\Query\CountDailyEventsByAppHandler'
         arguments:
             - '@akeneo_app.audit.persistence.query.select_apps_event_count_by_date'
