+++ src/Akeneo/Apps/back/tests/Unit/spec/Audit/Application/Query/CountDailyEventsByAppHandlerSpec.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace spec\Akeneo\Apps\Audit\Domain\Model\Read;
+namespace spec\Akeneo\Apps\Audit\Application\Query;
 
-use Akeneo\Apps\Audit\Application\Query\FetchAppsEventCountByEventHandler;
-use Akeneo\Apps\Audit\Application\Query\FetchAppsEventCountByEventQuery;
+use Akeneo\Apps\Audit\Application\Query\CountDailyEventsByAppHandler;
+use Akeneo\Apps\Audit\Application\Query\CountDailyEventsByAppQuery;
 use Akeneo\Apps\Audit\Domain\Model\Read\EventCountByApp;
 use Akeneo\Apps\Audit\Domain\Model\Read\EventCountByDate;
 use Akeneo\Apps\Audit\Domain\Persistence\Query\SelectAppsEventCountByDateQuery;
@@ -16,7 +16,7 @@ use PhpSpec\ObjectBehavior;
  * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
  * @license http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-class FetchAppsEventCountByEventHandlerSpec extends ObjectBehavior
+class CountDailyEventsByAppHandlerSpec extends ObjectBehavior
 {
     function let(SelectAppsEventCountByDateQuery $selectAppsEventCountByDateQuery)
     {
@@ -25,17 +25,21 @@ class FetchAppsEventCountByEventHandlerSpec extends ObjectBehavior
 
     function it_is_initializable()
     {
-        $this->shouldBeAnInstanceOf(FetchAppsEventCountByEventHandler::class);
+        $this->shouldBeAnInstanceOf(CountDailyEventsByAppHandler::class);
     }
 
     function it_handles_the_event_count($selectAppsEventCountByDateQuery)
     {
-        $eventCountByApp = new EventCountByApp('Magento');
-        $eventCountByApp->addEventCount(new EventCountByDate(42, new \DateTime('2019-12-10')));
+        $eventCountByApp1 = new EventCountByApp('Magento');
+        $eventCountByApp1->addEventCount(new EventCountByDate(42, new \DateTime('2019-12-10')));
+        $eventCountByApp1->addEventCount(new EventCountByDate(123, new \DateTime('2019-12-11')));
+
+        $eventCountByApp2 = new EventCountByApp('Bynder');
+        $eventCountByApp2->addEventCount(new EventCountByDate(36, new \DateTime('2019-12-11')));
 
         $selectAppsEventCountByDateQuery
             ->execute('product_created', '2019-12-10', '2019-12-12')
-            ->willReturn([$eventCountByApp]);
+            ->willReturn([$eventCountByApp1, $eventCountByApp2]);
 
         $expectedData = [
             'Magento' => [
@@ -46,7 +50,7 @@ class FetchAppsEventCountByEventHandlerSpec extends ObjectBehavior
                 '2019-12-11' => 36,
             ]
         ];
-        $query = new FetchAppsEventCountByEventQuery('product_created', '2019-12-12', '2019-12-14');
-        $this->handle($query)->shouldReturn($expectedData);
+        $query = new CountDailyEventsByAppQuery('product_created', '2019-12-10', '2019-12-12');
+        $this->handle($query)->shouldBeLike($expectedData);
     }
 }
